Show duration of flights as a float

+++ db/migrate/20251027192430_create_flights.rb
@@ -1,7 +1,7 @@
 class CreateFlights < ActiveRecord::Migration[8.0]
   def change
     create_table :flights do |t|
-      t.integer :duration
+      t.float :duration
       t.string :note
       t.references :flying_session, null: false, foreign_key: true
